Change file volumes to folders
If a volume is set to be a file, but it does not exist when docker
compose is started up, it is created as a directory. This causes issues
for those first starting up the docker containers, as some of the log
files are created as directories, instead of files as they should be.
The workaround is simply to use a directory volume, docker/logs/ instead
of a specific file.